+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 - [#174](https://github.com/os2display/display-templates/pull/174)
   - Fixed test issue with calendar dates.
+- [#161](https://github.com/os2display/display-templates/pull/161) 
+  - Added social news template. 
 
 ## [2.4.0] - 2025-03-31